refactor providers and add oracle cloudshell provider with support for delegation token
