+++ fn_go.go
@@ -6,6 +6,14 @@ import (
 	"github.com/fnproject/fn_go/provider/oracle"
 )
 
+const (
+	// provider names
+	DefaultProvider  = "default"
+	OracleProvider   = "oracle"
+	OracleIPProvider = "oracle-ip"
+	OracleCSProvider = "oracle-cs"
+)
+
 // DefaultProviders includes the bundled providers available in the client
 var DefaultProviders = provider.Providers{
 	Providers: map[string]provider.ProviderFunc{
@@ -13,5 +21,6 @@ var DefaultProviders = provider.Providers{
 		"default":   defaultprovider.NewFromConfig,
 		"oracle":    oracle.NewFromConfig,
 		"oracle-ip": oracle.NewIPProvider,
+		"oracle-cs": oracle.NewCSProvider,
 	},
 }
